% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPEtabSBML.R
\name{importPEtabSBML}
\alias{importPEtabSBML}
\title{Import an SBML model and corresponding PEtab objects}
\usage{
importPEtabSBML(
  modelname = "Boehm_JProteomeRes2014",
  path2BC = "BenchmarkModels/",
  path2TestCases = "PEtabTests/",
  TestCases = FALSE,
  compile = FALSE,
  SBML_file = NULL,
  observable_file = NULL,
  condition_file = NULL,
  data_file = NULL,
  parameter_file = NULL,
  assign_reactions = NULL,
  assign_observables = NULL,
  assign_g = NULL,
  assign_data = NULL,
  assign_ODEmodel = NULL,
  assign_condition.grid = NULL,
  assign_trafo = NULL,
  assign_p = NULL,
  assign_x = NULL,
  assign_pouter = NULL,
  assign_obj = NULL,
  assign_times = NULL,
  assign_err = NULL,
  assign_errors = NULL
)
}
\arguments{
\item{parameters}{PEtab parameter file as .tsv}
}
\value{
name of imported model
}
\description{
This function imports an SBML model and corresponding PEtab objects, e.g. from the Benchmark collection.
Note: Objects such as model equations, parameters or data are automatically assigned to standard variables (reactions, observables, g, x, p0) and written to your current working directory (via <<-).
You can obtain your common variable names by the additional assign arguments.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
