% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynormL2.R
\name{mynormL2}
\alias{mynormL2}
\title{Calculate LL of model}
\usage{
mynormL2(
  data,
  x,
  errmodel = NULL,
  times = NULL,
  attr.name = "data",
  outputLL = FALSE
)
}
\description{
This function calculates the likelihood of a model.
}
\author{
Marcus Rosenblatt and Svenja Kemmer
}
